rag/requirements.txt
# RAG Pipeline Requirements
# Core dependencies for LLM integration

# Existing dependencies from main requirements.txt
beautifulsoup4>=4.9.0
mysql-connector-python>=8.0.0
pysolr>=3.9.0

# New RAG-specific dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
ollama>=0.1.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Optional but recommended
python-multipart>=0.0.6  # For file uploads
httpx>=0.24.0           # For HTTP requests
python-jose[cryptography]>=3.3.0  # For JWT tokens
python-dotenv>=1.0.0    # For environment management
aiofiles>=23.0.0        # For async file operations